CmakeLists refactoring, added autoware compiler flags, minor changes in code

+++ src/system/autoware_state_monitor/src/autoware_state_monitor_node/state_machine.cpp
@@ -39,7 +39,7 @@ bool isStopped(
   const double th_stopped_velocity_mps)
 {
   for (const auto & odometry : odometry_buffer) {
-    if (std::abs(odometry->velocity_mps) > th_stopped_velocity_mps) {
+    if (std::abs(static_cast<double>(odometry->velocity_mps)) > th_stopped_velocity_mps) {
       return false;
     }
   }
